+++ CHANGES.md
@@ -3,6 +3,8 @@
 
 * Use custom classes instead of `Foldable`. Thanks to this, `length` and similar functions can't anymore be used on tuples or `Maybe`, but can be used on e.g. `Text`, `ByteString` and `IntSet`.
 
+* Add `allM`, `anyM,` `andM`, `orM`.
+
 0.1.11
 ======